# specs/005-interview-read-model/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: JobTrace Private Interview Read Model API
  version: 0.1.0
  description: Owner-scoped, read-only compatibility contract for feature 005; no production routing is enabled.
paths:
  /api/interviews:
    get:
      operationId: listPrivateInterviews
      summary: List the authenticated owner's interview reviews
      security:
        - signedIdentityBridge: []
      parameters:
        - { name: applicationId, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, schema: { type: string, maxLength: 200 } }
        - name: status
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/ReviewStatus' } }
        - name: stage
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/InterviewStage' } }
        - name: result
          in: query
          style: form
          explode: true
          schema: { type: array, items: { $ref: '#/components/schemas/RoundResult' } }
        - name: publication
          in: query
          style: form
          explode: true
          schema:
            type: array
            items: { type: string, enum: [private, anonymous, attributed] }
        - { name: interviewedFrom, in: query, schema: { type: string, format: date } }
        - { name: interviewedTo, in: query, schema: { type: string, format: date } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
      responses:
        '200':
          description: Owner-scoped review page, newest interview date first
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InterviewPage' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/interviews/{id}:
    get:
      operationId: getPrivateInterview
      summary: Return one review owned by the authenticated user
      security:
        - signedIdentityBridge: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Owner-scoped complete review
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InterviewDetail' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/applications/{id}/detail:
    get:
      operationId: getApplicationDialogData
      summary: Return an owned application detail with its private interview summaries
      security:
        - signedIdentityBridge: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Existing application detail plus owner-scoped interview summaries
          headers:
            Cache-Control:
              schema: { type: string, const: 'private, no-store' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplicationDialogData' }
        '400': { $ref: '#/components/responses/InvalidRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  securitySchemes:
    signedIdentityBridge:
      type: apiKey
      in: header
      name: Authorization
      description: Internal-only, request-bound signed identity assertion from feature 003.
  responses:
    InvalidRequest:
      description: Invalid filter, cursor, limit, or identifier
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: No trusted authenticated owner
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Record is missing or not owned by the authenticated user
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unavailable:
      description: Required storage or identity dependency is unavailable
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    InterviewStage:
      type: string
      enum: [assessment, interview_1, interview_2, interview_3, hr_interview, final_interview]
    ReviewStatus:
      type: string
      enum: [draft, pending_review, completed]
    RoundResult:
      type: string
      enum: [pending, passed, failed]
    ApplicationType:
      type: string
      enum: [summer_internship, daily_internship, spring_recruitment, early_campus_recruitment, campus_recruitment, social_recruitment]
    RecruitmentStage:
      type: string
      enum: [screening, assessment, written_test, interview_1, interview_2, interview_3, hr_interview, final_interview]
    InterviewSummary:
      type: object
      required: [id, applicationId, stageOccurrenceId, companyName, positionName, stage, interviewedOn, status, roundResult, linked, questionCount, actionCount, visibility, authorMode, publishedAt]
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        stageOccurrenceId: { type: [string, 'null'], format: uuid }
        companyName: { type: string }
        positionName: { type: string }
        stage: { $ref: '#/components/schemas/InterviewStage' }
        interviewedOn: { type: string, format: date }
        status: { $ref: '#/components/schemas/ReviewStatus' }
        roundResult: { $ref: '#/components/schemas/RoundResult' }
        linked: { type: boolean }
        questionCount: { type: integer, minimum: 0 }
        actionCount: { type: integer, minimum: 0 }
        visibility: { type: string, enum: [private, public] }
        authorMode: { type: string, enum: [anonymous, attributed] }
        publishedAt: { type: [string, 'null'] }
    InterviewPage:
      type: object
      additionalProperties: false
      required: [items, nextCursor, total, limit]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/InterviewSummary' } }
        nextCursor: { type: [string, 'null'] }
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1, maximum: 100 }
    InterviewQuestion:
      type: object
      additionalProperties: false
      required: [id, category, question, originalAnswer, followUpNotes, improvedAnswer, selfRating]
      properties:
        id: { type: string, format: uuid }
        category: { type: string, enum: [technical, project, behavioral, system_design, other] }
        question: { type: string }
        originalAnswer: { type: [string, 'null'] }
        followUpNotes: { type: [string, 'null'] }
        improvedAnswer: { type: [string, 'null'] }
        selfRating: { type: [integer, 'null'], minimum: 1, maximum: 5 }
    InterviewActionItem:
      type: object
      additionalProperties: false
      required: [id, content, completed]
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        completed: { type: boolean }
    InterviewDetail:
      allOf:
        - { $ref: '#/components/schemas/InterviewSummary' }
        - type: object
          required: [format, durationMinutes, interviewerNotes, highlights, gaps, version, questions, actionItems, createdAt, updatedAt]
          properties:
            format: { type: [string, 'null'], enum: [online, offline, phone, null] }
            durationMinutes: { type: [integer, 'null'], minimum: 1, maximum: 600 }
            interviewerNotes: { type: [string, 'null'] }
            highlights: { type: [string, 'null'] }
            gaps: { type: [string, 'null'] }
            version: { type: integer, minimum: 1 }
            questions: { type: array, items: { $ref: '#/components/schemas/InterviewQuestion' } }
            actionItems: { type: array, items: { $ref: '#/components/schemas/InterviewActionItem' } }
            createdAt: { type: string }
            updatedAt: { type: string }
    StageInterviewSummary:
      type: object
      additionalProperties: false
      required: [id, stage, interviewedOn, status, questionCount, stageOccurrenceId]
      properties:
        id: { type: string, format: uuid }
        stage: { $ref: '#/components/schemas/InterviewStage' }
        interviewedOn: { type: string, format: date }
        status: { $ref: '#/components/schemas/ReviewStatus' }
        questionCount: { type: integer, minimum: 0 }
        stageOccurrenceId: { type: [string, 'null'], format: uuid }
    ApplicationDialogData:
      type: object
      additionalProperties: false
      required: [application, interviews]
      properties:
        application: { $ref: '#/components/schemas/ApplicationDetail' }
        interviews: { type: array, items: { $ref: '#/components/schemas/StageInterviewSummary' } }
    ApplicationDetail:
      description: Same core application detail fields and semantics as feature 004.
      type: object
      additionalProperties: false
      required: [id, companyName, positionName, city, jobUrl, appliedDate, type, status, latestDate, stages, needsFollowUp, followUpDays, followUpReason, version, notes, stageOccurrences, events, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        companyName: { type: string }
        positionName: { type: string }
        city: { type: [string, 'null'] }
        jobUrl: { type: [string, 'null'] }
        appliedDate: { type: string, format: date }
        type: { $ref: '#/components/schemas/ApplicationType' }
        status: { type: string, enum: [submitted, offer, refused] }
        latestDate: { type: string, format: date }
        stages: { type: array, items: { $ref: '#/components/schemas/RecruitmentStage' } }
        needsFollowUp: { type: boolean }
        followUpDays: { type: integer, minimum: 0 }
        followUpReason: { type: [string, 'null'], enum: [timeline, application, null] }
        version: { type: integer, minimum: 1 }
        notes: { type: [string, 'null'] }
        stageOccurrences: { type: array, items: { $ref: '#/components/schemas/ApplicationStageOccurrence' } }
        events: { type: array, items: { $ref: '#/components/schemas/ApplicationEvent' } }
        createdAt: { type: string }
        updatedAt: { type: string }
    ApplicationStageOccurrence:
      type: object
      additionalProperties: false
      required: [id, stage, occurredOn]
      properties:
        id: { type: string, format: uuid }
        stage: { $ref: '#/components/schemas/RecruitmentStage' }
        occurredOn: { type: string, format: date }
    ApplicationEvent:
      type: object
      additionalProperties: false
      required: [id, type, occurredOn, before, after, createdAt]
      properties:
        id: { type: string, format: uuid }
        type: { type: string }
        occurredOn: { type: string, format: date }
        before: {}
        after: {}
        createdAt: { type: string }
    Problem:
      type: object
      required: [status, title, detail, code, requestId]
      properties:
        status: { type: integer }
        title: { type: string }
        detail: { type: string }
        code: { type: string }
        requestId: { type: string }